+++ benches/support/mod.rs
@@ -1,6 +1,6 @@
 #![allow(dead_code)]
 use core::f32;
-use glam::{Mat2, Mat3, Mat4, Quat, Vec2, Vec3, Vec3A, Vec4};
+use glam::{Affine3D, Mat2, Mat3, Mat4, Quat, Vec2, Vec3, Vec3A, Vec4};
 
 pub struct PCG32 {
     state: u64,
@@ -107,6 +107,15 @@ pub fn random_srt_mat3(rng: &mut PCG32) -> Mat3 {
         random_vec2(rng),
     )
 }
+
+pub fn random_srt_affine3d(rng: &mut PCG32) -> Affine3D {
+    Affine3D::from_scale_rotation_translation(
+        random_nonzero_vec3(rng),
+        random_quat(rng),
+        random_vec3(rng),
+    )
+}
+
 pub fn random_srt_mat4(rng: &mut PCG32) -> Mat4 {
     Mat4::from_scale_rotation_translation(
         random_nonzero_vec3(rng),