+++ benches/vec3.rs
@@ -6,7 +6,7 @@ mod support;
 use criterion::{criterion_group, criterion_main, Criterion};
 use glam::Vec3;
 use std::ops::{Add, Mul};
-use support::{random_mat3, random_quat, random_srt_mat4, random_vec3};
+use support::{random_mat3, random_quat, random_vec3};
 
 bench_binop!(
     vec3_add_vec3,
@@ -29,22 +29,6 @@ bench_binop!(
     "mat3 mul vec3",
     op => mul,
     from1 => random_mat3,
-    from2 => random_vec3
-);
-
-bench_binop!(
-    mat4_transform_point3,
-    "mat4 transform point3",
-    op => transform_point3,
-    from1 => random_srt_mat4,
-    from2 => random_vec3
-);
-
-bench_binop!(
-    mat4_transform_vector3,
-    "mat4 transform vector3",
-    op => transform_vector3,
-    from1 => random_srt_mat4,
     from2 => random_vec3
 );
 
@@ -184,8 +168,6 @@ bench_select!(
 criterion_group!(
     benches,
     mat3_mul_vec3,
-    mat4_transform_point3,
-    mat4_transform_vector3,
     quat_mul_vec3,
     vec3_add_vec3,
     vec3_angle_between,