+++ benches/vec4.rs
@@ -5,16 +5,7 @@ mod support;
 
 use criterion::{criterion_group, criterion_main, Criterion};
 use glam::Vec4;
-use std::ops::Mul;
-use support::{random_srt_mat4, random_vec4};
-
-bench_binop!(
-    vec4_mul_mat4,
-    "vec4 mul mat4",
-    op => mul,
-    from1 => random_srt_mat4,
-    from2 => random_vec4
-);
+use support::random_vec4;
 
 bench_select!(
     vec4_select,
@@ -24,6 +15,6 @@ bench_select!(
     from => random_vec4
 );
 
-criterion_group!(benches, vec4_mul_mat4, vec4_select,);
+criterion_group!(benches, vec4_select);
 
 criterion_main!(benches);