+++ tests/mat4.rs
@@ -302,6 +302,17 @@ macro_rules! impl_mat4_tests {
             assert_approx_eq!($mat4::IDENTITY, m * m_inv, 1.0e-5);
             assert_approx_eq!($mat4::IDENTITY, m_inv * m, 1.0e-5);
             assert_approx_eq!(m_inv, trans_inv * rotz_inv * scale_inv, 1.0e-6);
+
+            // Make sure we can invert a "random" matrix:
+            let m = $mat4::from_cols(
+                $newvec4(1.0, -0.3, 1.0, 1.0),
+                $newvec4(0.5, 0.6, 0.7, 0.8),
+                $newvec4(-0.9, -0.3, 0.0, 12.0),
+                $newvec4(0.13, 0.14, 0.15, 0.16),
+            );
+            let m_inv = m.inverse();
+            assert_approx_eq!($mat4::IDENTITY, m * m_inv, 1.0e-5);
+            assert_approx_eq!($mat4::IDENTITY, m_inv * m, 1.0e-5);
         }
 
         #[test]