+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog], and this project adheres to
 [Semantic Versioning].
 
+## Unreleased
+
+### Added
+
+* Added `Affine3D`: an affine 3D transform.
+
 ## [0.14.0] - 2021-04-09
 
 ### Breaking changes
@@ -20,7 +26,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 ### Changed
 
-* Updated dependency versions of `bytemuck` to 1.5, `rand` to 0.8, 
+* Updated dependency versions of `bytemuck` to 1.5, `rand` to 0.8,
   `rand_xoshiro` to 0.6 and `spriv-std` to 0.4.0-alpha4.
 
 ## [0.13.1] - 2021-03-24